+++ packages/core/docs/commercetools/changelog.md
@@ -8,6 +8,4 @@
 ### Changes
 
 * **factories**: refactored setup using apiClientFactory (#4777)
-* **factories**: updated return signature of loadCart ([#4616](https://github.com/DivanteLtd/vue-storefront/issues/4616))
-* **factories**: updated call arguments of applyCoupon ([#4616](https://github.com/DivanteLtd/vue-storefront/issues/4616))
-* **factories**: updated call arguments of removeCoupon ([#4616](https://github.com/DivanteLtd/vue-storefront/issues/4616))
+* **factories**: updated return signature of loadCart, updated call arguments of applyCoupon, removeCoupon, changed coupon to currentCoupon in singatures ([#4616](https://github.com/DivanteLtd/vue-storefront/issues/4616))